Refactor limit documents into save-image validation
