+++ apps/evisa/index.js
@@ -4,7 +4,6 @@
 // const config = require('../../config');
 const SaveImage = require('./behaviours/save-image');
 const RemoveImage = require('./behaviours/remove-image');
-const LimitDocument = require('./behaviours/limit-documents');
 //const agentEmail = require('./behaviours/agent-email')(config.email);
 
 module.exports = {
@@ -51,7 +50,7 @@ module.exports = {
       next: '/upload',
     },
     '/upload': {
-      behaviours: [SaveImage('image-upload'), RemoveImage, LimitDocument],
+      behaviours: [SaveImage('image-upload'), RemoveImage],
       fields: ['image-upload'],
       next: '/confirmation',
     },
