+++ apps/evisa/translations/src/en/pages.json
@@ -30,7 +30,7 @@
     "upload-image": "Upload an image",
     "no-files-uploaded": "No files uploaded",
     "files-uploaded": "Files uploaded",
-    "upload-maximum": "You may upload a maximum of {{maxUploads}} files"
+    "upload-maximum": "You may submit a maximum of {{maxUploads}} files"
   },
   "confirmation": {
     "header": "Question sent",
