+++ apps/evisa/translations/src/en/validation.json
@@ -49,5 +49,13 @@
     "minlength": "Enter your question",
     "maxlength": "Your question must be 2,000 characters or less",
     "invalidCharactersValidator": "Your question must not include these characters: [ ] < > / |"
+  },
+  "image-upload": {
+    "maxFileSize": "The selected file must be 25MB or less",
+    "fileType": "The selected file must be JPG, PNG, GIF",
+    "default": "The selected file was not uploaded - try again",
+    "unsaved": "The selected file was not uploaded - try again",
+    "tooMany": "You can only submit up to 5 files.",
+    "virusCheck": "The file contains a virus"
   }
 }
